Move `mark_as_*` methods to `cfg.rs`
