+++ cobalt-ast/src/lib.rs
@@ -4,7 +4,6 @@ pub mod cfg;
 pub mod context;
 pub mod dottedname;
 pub mod error;
-pub mod ops;
 pub mod serial_utils;
 pub mod types;
 pub mod value;
